First take on better message routing.
Three go routines:
* Registry keeper, knows how to route messages to results
* Outbound writer, serializes message
* Inbound reader, deserializes message and dispatches to registry

An API call registers its context into the registry and posts the
message to the outbound queue. Once the inbound reader has got the
reply it dispatches it to the registry which looks up the cookie
and dispatches the value to the callee.

Everything is a mess, next will be to find better places for all
of this with a clean API.