Add a RW-lock to facilitate shutdown of command channel.
Requests grab a RLock when posting to the command channel, and if
the command channel is nil, the client isn't connected and thus an
error message is posted back immediately.

When shutting down, an RWLock is grabbed, at the same time as looping
over the contents of the command channel. If the RWLock hasn't been
locked yet, the loop will block once the last producer has posted its
request, at which point the RWLock grabbing will succeed and the channel
closed and set to nil which will allow the loop to exit. As the channel
is now nil, the interaction is back to the first paragraph here.

The same RWLock will be grabbed during Dial when setting up the channels.

Still some resource management left to deal with, but this might be an
ok strategy for the time being.