Refactor method/broadcast parsing in genipc.
Broadcasts were actually broken as their Id's are generated
in the same namespace as Signals. By pure luck the Broadcasts
tested were of lower Id than any Signal, so they happened to work.

Now Signals and Broadcasts are collected in the same slice and
distinguished by a ResultClass enum that is used to filter out
what is being generated for a given go generate call.